[WIP] Add an option to use nominal types
Add an experimental option to enable nominal types from the command line.
When parsing the text format, subtype relations are inferred from globals of the
form

```
  (global (rtt ...) (rtt.sub $<subtype> (global.get $<supertype>.rtt)))
```

Unfortunately, the binary parser is not set up to make it feasible to do the
same trick when parsing binaries. Also, J2CL output currently does not not emit
rtt globals encoding vtable subtyping, so this hack is not sufficient for
real-world programs.

This PR is for informational purposes only and still contains debugging prints
that are useful for debugging why subtyping fails to validate.